Issue/644: refactor
