+++ presentation/pixi/pixi-preloader.mjs
@@ -3,66 +3,123 @@ import { FOUNDRY_10_PIXI_VERSION, PIXI_VERSION } from "./pixi-globals.mjs";
 const BASE_PATH = "systems/strive/presentation/image";
 
 /**
- * Constants of all the image texture urls for use with PixiJS.
+ * Wraps PixiJs functionality for ease of use. 
  * 
- * @constant
- * 
- * @property {String} ACTION_POINT_EMPTY 
- * @property {String} ACTION_POINT_FULL 
- * @property {String} CARET_LEFT 
- * @property {String} CARET_RIGHT 
- */
-export const TEXTURES = {
-  ACTION_POINT_EMPTY: `${BASE_PATH}/action-point-empty.png`,
-  ACTION_POINT_FULL: `${BASE_PATH}/action-point-full.png`,
-  CARET_LEFT: `${BASE_PATH}/caret-left-solid.png`,
-  CARET_RIGHT: `${BASE_PATH}/caret-right-solid.png`,
-};
-
-/**
- * A map of `TEXTURES` mapped to a texture instance that was 
- * preloaded. 
- * @type {Map<String, PIXI.Texture>}
- */
-const preloadedTextures = new Map();
-
-/**
- * Preloads all custom PIXI textures and caches them for synchronous access later. 
- * 
- * @async
- */
-export async function preloadPixiTextures() {
-  if (PIXI_VERSION.greaterThan(FOUNDRY_10_PIXI_VERSION)) {
-    for (const propertyName in TEXTURES) {
-      const url = TEXTURES[propertyName];
-      const texture = await PIXI.Assets.load(url);
-      preloadedTextures.set(url, texture);
-    }
-  } else {
-    await PIXI.Loader.shared
-      .add(TEXTURES.ACTION_POINT_EMPTY)
-      .add(TEXTURES.ACTION_POINT_FULL)
-      .add(TEXTURES.CARET_LEFT)
-      .add(TEXTURES.CARET_RIGHT)
-      .load();
-  }
-}
-
-/**
- * Fetches and returns the PIXI texture identified by the given key. 
- * 
- * Abstracts away the exact access mechanism, so that changes in PIXI's API 
+ * Abstracts away exact access mechanisms, so that changes in PIXI's API 
  * only impact **this** piece of code. 
  * 
- * @param {String} key One of the `TEXTURES` constants that identify 
- * one of the textures. 
- * 
- * @returns {Object}
+ * @constant
  */
-export function getPixiTexture(key) {
-  if (PIXI_VERSION.greaterThan(FOUNDRY_10_PIXI_VERSION)) {
-    return preloadedTextures.get(key);
-  } else {
-    return PIXI.Loader.shared.resources[key].texture;
+export const PixiLoader = {
+  /**
+   * Constants of all the image texture urls for use with PixiJS.
+   * 
+   * @constant
+   * 
+   * @property {String} ACTION_POINT_EMPTY 
+   * @property {String} ACTION_POINT_FULL 
+   * @property {String} CARET_LEFT 
+   * @property {String} CARET_RIGHT 
+   */
+  TEXTURES: {
+    ACTION_POINT_EMPTY: `${BASE_PATH}/action-point-empty.png`,
+    ACTION_POINT_FULL: `${BASE_PATH}/action-point-full.png`,
+    CARET_LEFT: `${BASE_PATH}/caret-left-solid.png`,
+    CARET_RIGHT: `${BASE_PATH}/caret-right-solid.png`,
+  },
+
+  /**
+   * Constants of all the vector graphics urls for use with PixiJS.
+   * 
+   * @constant
+   * 
+   * @property {String} ACTION_POINT_EMPTY 
+   * @property {String} ACTION_POINT_FULL 
+   * @property {String} CARET_LEFT 
+   * @property {String} CARET_RIGHT 
+   */
+  VECTOR_GRAPHICS: {
+    // CARET_LEFT: `${BASE_PATH}/caret-left-solid.svg`,
+    // CARET_RIGHT: `${BASE_PATH}/caret-right-solid.svg`,
+  },
+
+  /**
+   * A map of `TEXTURES` mapped to a texture instance that was 
+   * preloaded. 
+   * @type {Map<String, PIXI.Texture>}
+   */
+  _preloadedTextures: new Map(),
+
+
+  /**
+   * A map of `VECTOR_GRAPHICS` mapped to a graphics instance that was 
+   * preloaded. 
+   * @type {Map<String, Graphics>}
+   */
+  _preloadedGraphics: new Map(),
+
+  /**
+   * Preloads all custom PIXI textures and caches them for synchronous access later. 
+   * 
+   * @async
+   */
+  preloadTextures: async function () {
+    if (PIXI_VERSION.greaterThan(FOUNDRY_10_PIXI_VERSION)) {
+      for (const propertyName in this.TEXTURES) {
+        const url = this.TEXTURES[propertyName];
+        const texture = await PIXI.Assets.load(url);
+        this._preloadedTextures.set(url, texture);
+      }
+    } else {
+      await PIXI.Loader.shared
+        .add(this.TEXTURES.ACTION_POINT_EMPTY)
+        .add(this.TEXTURES.ACTION_POINT_FULL)
+        .add(this.TEXTURES.CARET_LEFT)
+        .add(this.TEXTURES.CARET_RIGHT)
+        .load();
+    }
+  },
+
+  /**
+   * Preloads all custom PIXI graphics and caches them for synchronous access later. 
+   * 
+   * @async
+   */
+  preloadGraphics: async function () {
+    for (const propertyName in this.VECTOR_GRAPHICS) {
+      const url = this.VECTOR_GRAPHICS[propertyName];
+      const svg = await PIXI.Assets.load(url, {
+        parseAsGraphicsContext: true, // If false, it returns a texture instead.
+      });
+      const graphics = new Graphics(svg);
+      this._preloadedGraphics.set(url, graphics);
+    }
+  },
+
+  /**
+   * Fetches and returns the PIXI texture identified by the given key. 
+   * 
+   * @param {String} key One of the `TEXTURES` constants. 
+   * 
+   * @returns {PIXI.Texture}
+   */
+  getTexture: function (key) {
+    if (PIXI_VERSION.greaterThan(FOUNDRY_10_PIXI_VERSION)) {
+      return this._preloadedTextures.get(key);
+    } else {
+      return PIXI.Loader.shared.resources[key].texture;
+    }
+  },
+
+  /**
+   * Fetches and returns the PIXI graphics instance identified by the given key. 
+   * 
+   * @param {String} key One of the `VECTOR_GRAPHICS` constants. 
+   * 
+   * @returns {PIXI.Texture}
+   */
+  getGraphics: function (key) {
+    return this._preloadedGraphics.get(key);
   }
-}
+};
+
